+++ src/dataspec/base.py
@@ -65,6 +65,7 @@ SpecPredicate = Union[  # type: ignore
     List["SpecPredicate"],  # type: ignore
     FrozenSet[Any],
     Set[Any],
+    Type[Any],
     PredicateFn,
     ValidatorFn,
     "Spec",
@@ -381,37 +382,22 @@ class ValidatorSpec(Spec):
 
     @classmethod
     def from_validators(
-        cls,
-        tag: Tag,
-        *preds: Union[ValidatorFn, "ValidatorSpec"],
-        conformer: Optional[Conformer] = None,
+        cls, tag: Tag, *preds: ValidatorFn, conformer: Optional[Conformer] = None,
     ) -> Spec:
         """Return a single Validator spec from the composition of multiple validator
-        functions or ValidatorSpec instances."""
+        functions."""
         assert len(preds) > 0, "At least on predicate must be specified"
 
-        # Avoid wrapping an existing validator spec or singleton spec in an extra layer
-        # of indirection
+        # Avoid wrapping an existing validator function in an extra layer of
+        # indirection
         if len(preds) == 1:
-            if isinstance(preds[0], ValidatorSpec):
-                return preds[0]
-            else:
-                return ValidatorSpec(tag, preds[0], conformer=conformer)
-
-        specs = []
-        for pred in preds:
-            if isinstance(pred, ValidatorSpec):
-                specs.append(pred)
-            else:
-                specs.append(ValidatorSpec(pred.__name__, pred))
+            return ValidatorSpec(tag, preds[0], conformer=conformer)
 
         def do_validate(v) -> Iterator[ErrorDetails]:
-            for spec in specs:
-                yield from spec.validate(v)
+            for pred in preds:
+                yield from pred(v)
 
-        return cls(
-            tag, do_validate, compose_spec_conformers(*specs, conform_final=conformer)
-        )
+        return cls(tag, do_validate, conformer=conformer)
 
 
 @attr.s(auto_attribs=True, frozen=True, slots=True)
@@ -571,7 +557,7 @@ class CollSpec(Spec):
         return cls(
             tag or "coll",
             spec=spec,
-            conformer=compose_conformers(conform_coll, *filter(None, (conformer,))),
+            conformer=compose_conformers(conform_coll, conformer),
             out_type=out_type,
             validate_coll=validate_coll,
         )
@@ -634,7 +620,7 @@ class DictSpec(Spec):
             tag or "map",
             reqkeyspecs=reqkeys,
             optkeyspecs=optkeys,
-            conformer=compose_conformers(conform_mapping, *filter(None, (conformer,))),
+            conformer=compose_conformers(conform_mapping, conformer),
         )
 
     def validate(self, d) -> Iterator[ErrorDetails]:  # pylint: disable=arguments-differ
@@ -650,7 +636,7 @@ class DictSpec(Spec):
                         via=[self.tag],
                         path=[k],
                     )
-        except TypeError:
+        except (AttributeError, TypeError):
             yield ErrorDetails(
                 message="Value is not a mapping type",
                 pred=self,
@@ -686,6 +672,81 @@ class DictSpec(Spec):
             {k: all_spec(str(k), *v) for k, v in map_pred.items()},
             conformer=conformer,
         )
+
+
+def kv_spec(
+    tag_or_pred: Union[Tag, SpecPredicate],
+    *preds: SpecPredicate,
+    conform_keys: bool = False,
+    conformer: Optional[Conformer] = None,
+) -> Spec:
+    """
+    Return a Spec that validates mapping types against a single Spec for all keys
+    and a single Spec for all values.
+
+    If ``conform_keys`` is specified as :py:obj:`True`, the default conformer will
+    conform keys and values. By default, ``conform_keys`` is :py:obj:`False` to avoid
+    duplicate names produced during the conformation.
+
+    The returned Spec's :py:meth:`dataspec.Spec.conform` method will return a
+    :py:class:`dict` with values conformed by the corresponding input Spec. If a
+    ``conformer`` is provided via keyword argument, that conformer will be provided a
+    :py:class:`dict` with the conformed :py:class:`dict` as described above. Otherwise,
+    the default conformer will simply return the conformed :py:class:`dict` . Note that
+    the default conformer does **not** modify the input mapping in place.
+
+    Exactly two Specs must be provided or a :py:class:`ValueError` will be raised
+    during construction.
+
+    :param tag_or_pred: an optional tag for the resulting spec or the key Spec
+        or value which can be converted into a Spec
+    :param preds: if a tag is given, preds should be exactly two Specs or values
+        which can be converted into Specs; the first shall be the Spec for the
+        keys and the second shall be the Spec for values
+    :param conform_keys: if :py:obj:`True`, the default conformer will also conform
+        keys according to the input key Spec; default is :py:obj:`False`
+    :param conformer: an optional conformer which will be composed with the
+        default conformer
+    :return: a Spec
+    """
+    tag, preds = tag_maybe(tag_or_pred, *preds)
+
+    if len(preds) != 2:
+        raise ValueError("Must specify a key and value Spec for k/v Specs")
+
+    tag = tag or "kv"
+    keyspec = make_spec(preds[0])
+    valspec = make_spec(preds[1])
+
+    def _kv_valid(d) -> Iterator[ErrorDetails]:
+        assert tag is not None
+
+        try:
+            for k, v in d.items():
+                yield from _enrich_errors(keyspec.validate(k), tag, d)
+                yield from _enrich_errors(valspec.validate(v), tag, k)
+        except (AttributeError, TypeError):
+            yield ErrorDetails(
+                message="Value is not a mapping type",
+                pred=_kv_valid,
+                value=d,
+                via=[tag],
+            )
+            return
+
+    if conform_keys:
+
+        def conform_mapping(d: Mapping) -> Mapping:
+            return {keyspec.conform(k): valspec.conform(v) for k, v in d.items()}
+
+    else:
+
+        def conform_mapping(d: Mapping) -> Mapping:
+            return {k: valspec.conform(v) for k, v in d.items()}
+
+    return ValidatorSpec(
+        tag, _kv_valid, conformer=compose_conformers(conform_mapping, conformer),
+    )
 
 
 @attr.s(auto_attribs=True, frozen=True, slots=True)
@@ -785,9 +846,7 @@ class SetSpec(Spec):
                     [mem, mem.name, mem.value] for mem in pred  # type: ignore[var-annotated]  # noqa
                 )
             ),
-            conformer=compose_conformers(
-                _enum_conformer(pred), *filter(None, (conformer,)),
-            ),
+            conformer=compose_conformers(_enum_conformer(pred), conformer,),
         )
 
 
@@ -828,7 +887,7 @@ class TupleSpec(Spec):
             tag or "tuple",
             pred=pred,
             specs=specs,
-            conformer=compose_conformers(conform_tuple, *filter(None, (conformer,))),
+            conformer=compose_conformers(conform_tuple, conformer),
             namedtuple=namedtuple_type,  # type: ignore
         )
 
@@ -866,12 +925,17 @@ def _enrich_errors(
         yield error.with_details(tag, loc=loc)
 
 
-def compose_conformers(*conformers: Conformer) -> Conformer:
+def compose_conformers(*conformers: Optional[Conformer]) -> Optional[Conformer]:
     """
     Return a single conformer which is the composition of the input conformers.
 
     If a single conformer is given, return the conformer.
     """
+
+    conformers = tuple(filter(None, conformers))
+
+    if not conformers:
+        return None
 
     if len(conformers) == 1:
         return conformers[0]
@@ -885,25 +949,6 @@ def compose_conformers(*conformers: Conformer) -> Conformer:
         return conformed_v
 
     return do_conform
-
-
-def compose_spec_conformers(
-    *specs: Spec, conform_final: Optional[Conformer] = None
-) -> Conformer:
-    """
-    Return a single conformer which is the composition of the conformers from each of
-    the child specs.
-
-    Apply the ``conform_final`` conformer on the final return from the composition, if
-    any.
-
-    This function bypasses the :py:meth:`dataspec.Spec.conform` method and accesses
-    :py:attr:`dataspec.Spec.conformer` directly.
-    """
-
-    return compose_conformers(
-        *filter(None, chain((spec.conformer for spec in specs), (conform_final,)))
-    )
 
 
 def all_spec(
@@ -935,8 +980,9 @@ def all_spec(
     This method is not suitable for producing a union of mapping Specs. To merge
     mapping Specs, use :py:meth:`dataspec.SpecAPI.merge` instead.
 
-    :param tag_or_pred: an optional tag for the resulting spec or the first Spec
-        or value which can be converted into a Spec
+    :param tag_or_pred: an optional tag for the resulting spec or the first Spec or
+        value which can be converted into a Spec; if no tag is provided, the default is
+        ``"all"``
     :param preds: zero or more Specs or values which can be converted into a Spec
     :param conformer: an optional conformer which will be applied to the final
         conformed value produced by the input Specs conformers
@@ -948,7 +994,11 @@ def all_spec(
         raise ValueError("Must provide at least one Spec for 'all' Specs")
 
     if len(preds) == 1:
-        return make_spec(*filter(None, (tag,)), preds[0], conformer=conformer)
+        return make_spec(
+            *filter(None, (tag,)),  # type: ignore[arg-type]  # noqa: F821
+            preds[0],
+            conformer=conformer,
+        )
 
     specs = [make_spec(pred) for pred in preds]
 
@@ -967,7 +1017,7 @@ def all_spec(
     return ValidatorSpec(
         tag or "all",
         _all_valid,
-        conformer=compose_spec_conformers(*specs, conform_final=conformer),
+        conformer=compose_conformers(*(spec.conformer for spec in specs), conformer,),
     )
 
 
@@ -1005,8 +1055,9 @@ def any_spec(
     :py:func:`dataspec.s` with the given ``tag`` and ``conformer`` and the value
     returned without merging.
 
-    :param tag_or_pred: an optional tag for the resulting spec or the first Spec
-        or value which can be converted into a Spec
+    :param tag_or_pred: an optional tag for the resulting spec or the first Spec or
+        value which can be converted into a Spec; if no tag is provided, the default is
+        ``"any"``
     :param preds: zero or more Specs or values which can be converted into a Spec
     :param tag_conformed: if :py:obj:`True`, the conformed value will be wrapped in a
         2-tuple where the first element is the successful spec and the second element
@@ -1020,7 +1071,11 @@ def any_spec(
         raise ValueError("Must provide at least one Spec for 'any' Specs")
 
     if len(preds) == 1:
-        return make_spec(*filter(None, (tag,)), preds[0], conformer=conformer)
+        return make_spec(
+            *filter(None, (tag,)),  # type: ignore[arg-type]  # noqa: F821
+            preds[0],
+            conformer=conformer,
+        )
 
     specs = [make_spec(pred) for pred in preds]
 
@@ -1082,8 +1137,9 @@ def merge_spec(
     input Specs in the order they were provided. Values with exactly one Spec will
     use that Spec as given.
 
-    :param tag_or_pred: an optional tag for the resulting spec or the first Spec
-        or value which can be converted into a Spec
+    :param tag_or_pred: an optional tag for the resulting spec or the first Spec or
+        value which can be converted into a Spec; if no tag is provided, the default is
+        computed as ``"merge-of-spec1-and-spec2-..."``
     :param preds: zero or more mapping Specs or values which can be converted into a
         mapping Spec
     :param conformer: an optional conformer for the value
@@ -1095,7 +1151,11 @@ def merge_spec(
         raise ValueError("Must provide at least one mapping Spec to merge")
 
     if len(preds) == 1:
-        return make_spec(*filter(None, (tag,)), preds[0], conformer=conformer)
+        return make_spec(
+            *filter(None, (tag,)),  # type: ignore[arg-type]  # noqa: F821
+            preds[0],
+            conformer=conformer,
+        )
 
     specs = [make_spec(pred) for pred in preds]
 
@@ -1191,7 +1251,9 @@ def type_spec(
 
 
 def make_spec(  # pylint: disable=inconsistent-return-statements  # noqa: MC0001
-    *args: Union[Tag, SpecPredicate], conformer: Optional[Conformer] = None
+    tag_or_pred: Union[Tag, SpecPredicate],
+    *preds: SpecPredicate,
+    conformer: Optional[Conformer] = None,
 ) -> Spec:
     """
     Create a new Spec instance from a :py:obj:`dataspec.base.SpecPredicate`.
@@ -1210,7 +1272,8 @@ def make_spec(  # pylint: disable=inconsistent-return-statements  # noqa: MC0001
     Specs may be created for enumerated types using a Python ``set`` or ``frozenset``
     or using Python :py:class:`enum.Enum` types. Specs created for enumerated types
     based on :py:class:`enum.Enum` values validate the Enum name, value, or Enum
-    singleton and conform the the Enum singleton.
+    singleton and conform the input value to the corresponding :py:class:`enum.Enum`
+    value.
 
     Specs may be created for homogeneous collections using a Python ``list`` type.
     Callers can specify a few additional parameters for collection specs by providing
@@ -1243,17 +1306,38 @@ def make_spec(  # pylint: disable=inconsistent-return-statements  # noqa: MC0001
     If both a new tag and conformer are given, a new Spec will be returned with both
     the new tag and conformer.
 
-    :param tag: an optional :py:data:`dataspec.Tag` for the resulting spec
-    :param pred: a value which can be be converted into a :py:class:`dataspec.Spec`
+    :param tag_or_pred: an optional tag for the resulting spec or a Spec or value which
+        can be converted into a Spec; if no tag is provided, the default depends on the
+        input type:
+
+        * for ``frozenset`` and ``set`` predicates, the default is ``"set"``
+
+        * for ``Enum`` predicates, the default is the name of the enum
+
+        * for ``tuple`` predicates, the default is ``"tuple"``
+
+        * for ``list`` (collection) predicates, the default is ``"coll"``
+
+        * for ``dict`` (mapping) predicates, the default is ``"map"``
+
+        * for ``type`` predicates, the default is the name of the type
+
+        * for callable predicates, the default is the name of the function
+
+    :param preds: if a tag is given, exactly one spec predicate; if no tag is given,
+        this should not be specified
     :param conformer: an optional :py:data:`dataspec.Conformer` for the value
     :return: a :py:class:`dataspec.base.Spec` instance
     """
-    tag = args[0] if isinstance(args[0], str) else None
+    tag, preds = tag_maybe(tag_or_pred, *preds)
+
+    if len(preds) != 1:
+        raise TypeError("Expected some spec predicate")
 
     try:
-        pred = args[0] if tag is None else args[1]
+        pred = preds[0]
     except IndexError:
-        raise TypeError("Expected some spec predicate; received only a Tag")
+        raise TypeError("Expected some spec predicate")
 
     if isinstance(pred, (frozenset, set)):
         return SetSpec(tag or "set", pred, conformer=conformer)